Tela de pesquisa item; Está consultando no banco de dados, só falta apresentar na tela.

+++ ClassLibrary/WebApplication/Documentos/Querys/Funcao CalculoDistancia.sql
@@ -11,8 +11,10 @@ create function CalculoDistancia(
 returns real
 as
 begin 
-	return sqrt( power(convert(bigint, @LatitudeVendedor) - convert(bigint, @LatitudeComprador), 2) + power( convert(bigint, @LongitudeVendedor) -  convert(bigint, @LongitudeComprador), 2))
+	return sqrt( power(convert(float, @LatitudeVendedor) - convert(float, @LatitudeComprador), 2) + power( convert(float, @LongitudeVendedor) -  convert(float, @LongitudeComprador), 2))
 end
+
+
